concurrent: Move tests
