+++ Orange/widgets/utils/concurrent.py
@@ -580,222 +580,3 @@ class methodinvoke(object):
         args = [Q_ARG(atype, arg) for atype, arg in zip(self.arg_types, args)]
         return QMetaObject.invokeMethod(
             self.obj, self.method, self.conntype, *args)
-
-
-import unittest
-
-
-class TestFutures(unittest.TestCase):
-    def test_futures(self):
-        f = Future()
-        self.assertEqual(f.done(), False)
-        self.assertEqual(f.running(), False)
-
-        self.assertTrue(f.cancel())
-        self.assertTrue(f.cancelled())
-
-        with self.assertRaises(CancelledError):
-            f.result()
-
-        with self.assertRaises(CancelledError):
-            f.exception()
-
-        f = Future()
-        f.set_running_or_notify_cancel()
-
-        with self.assertRaises(TimeoutError):
-            f.result(0.1)
-
-        with self.assertRaises(TimeoutError):
-            f.exception(0.1)
-
-        f = Future()
-        f.set_running_or_notify_cancel()
-        f.set_result("result")
-
-        self.assertEqual(f.result(), "result")
-        self.assertEqual(f.exception(), None)
-
-        f = Future()
-        f.set_running_or_notify_cancel()
-
-        f.set_exception(Exception("foo"))
-
-        with self.assertRaises(Exception):
-            f.result()
-
-        class Ref():
-            def __init__(self, ref):
-                self.ref = ref
-
-            def set(self, ref):
-                self.ref = ref
-
-        # Test that done callbacks are called.
-        called = Ref(False)
-        f = Future()
-        f.add_done_callback(lambda f: called.set(True))
-        f.set_result(None)
-        self.assertTrue(called.ref)
-
-        # Test that callbacks are called when cancelled.
-        called = Ref(False)
-        f = Future()
-        f.add_done_callback(lambda f: called.set(True))
-        f.cancel()
-        self.assertTrue(called.ref)
-
-        # Test that callbacks are called immediately when the future is
-        # already done.
-        called = Ref(False)
-        f = Future()
-        f.set_result(None)
-        f.add_done_callback(lambda f: called.set(True))
-        self.assertTrue(called.ref)
-
-        count = Ref(0)
-        f = Future()
-        f.add_done_callback(lambda f: count.set(count.ref + 1))
-        f.add_done_callback(lambda f: count.set(count.ref + 1))
-        f.set_result(None)
-        self.assertEqual(count.ref, 2)
-
-        # Test that the callbacks are called with the future as argument.
-        done_future = Ref(None)
-        f = Future()
-        f.add_done_callback(lambda f: done_future.set(f))
-        f.set_result(None)
-        self.assertIs(f, done_future.ref)
-
-
-class TestExecutor(unittest.TestCase):
-    def setUp(self):
-        self.app = QCoreApplication([])
-
-    def test_executor(self):
-        executor = ThreadExecutor()
-        f1 = executor.submit(pow, 100, 100)
-
-        f2 = executor.submit(lambda: 1 / 0)
-
-        f3 = executor.submit(QThread.currentThread)
-
-        self.assertTrue(f1.result(), pow(100, 100))
-
-        with self.assertRaises(ZeroDivisionError):
-            f2.result()
-
-        self.assertIsInstance(f2.exception(), ZeroDivisionError)
-
-        self.assertIsNot(f3.result(), QThread.currentThread())
-
-    def test_methodinvoke(self):
-        executor = ThreadExecutor()
-        state = [None, None]
-
-        class StateSetter(QObject):
-            @Slot(object)
-            def set_state(self, value):
-                state[0] = value
-                state[1] = QThread.currentThread()
-
-        def func(callback):
-            callback(QThread.currentThread())
-
-        obj = StateSetter()
-        f1 = executor.submit(func, methodinvoke(obj, "set_state", (object,)))
-        f1.result()
-
-        # So invoked method can be called
-        QCoreApplication.processEvents()
-
-        self.assertIs(state[1], QThread.currentThread(),
-                      "set_state was called from the wrong thread")
-
-        self.assertIsNot(state[0], QThread.currentThread(),
-                         "set_state was invoked in the main thread")
-
-        executor.shutdown(wait=True)
-
-    def test_executor_map(self):
-        executor = ThreadExecutor()
-
-        r = executor.map(pow, list(range(1000)), list(range(1000)))
-
-        results = list(r)
-
-        self.assertTrue(len(results) == 1000)
-
-
-class TestFutureWatcher(unittest.TestCase):
-    def setUp(self):
-        self.app = QCoreApplication([])
-
-    def test_watcher(self):
-        executor = ThreadExecutor()
-        f = executor.submit(QThread.currentThread)
-        watcher = FutureWatcher(f)
-
-        if f.cancel():
-            self.assertTrue(watcher.isCancelled())
-
-        executor.shutdown()
-
-
-class TestTask(unittest.TestCase):
-    def setUp(self):
-        self.app = QCoreApplication([])
-
-    def test_task(self):
-        results = []
-
-        task = Task(function=QThread.currentThread)
-        task.resultReady.connect(results.append)
-
-        task.start()
-        self.app.processEvents()
-
-        self.assertSequenceEqual(results, [QThread.currentThread()])
-
-        results = []
-
-        thread = QThread()
-        thread.start()
-
-        task = Task(function=QThread.currentThread)
-
-        task.moveToThread(thread)
-
-        self.assertIsNot(task.thread(), QThread.currentThread())
-        self.assertIs(task.thread(), thread)
-
-        task.resultReady.connect(results.append, Qt.DirectConnection)
-        task.start()
-
-        f = task.future()
-
-        self.assertIsNot(f.result(3), QThread.currentThread())
-
-        self.assertIs(f.result(3), results[-1])
-
-    def test_executor(self):
-        executor = ThreadExecutor()
-
-        f = executor.submit(QThread.currentThread)
-
-        self.assertIsNot(f.result(3), QThread.currentThread())
-
-        f = executor.submit(lambda: 1 / 0)
-
-        with self.assertRaises(ZeroDivisionError):
-            f.result()
-
-        results = []
-        task = Task(function=QThread.currentThread)
-        task.resultReady.connect(results.append, Qt.DirectConnection)
-
-        f = executor.submit_task(task)
-
-        self.assertIsNot(f.result(3), QThread.currentThread())
-
-        executor.shutdown()
